% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virion_class.R
\name{deposit}
\alias{deposit}
\title{deposit}
\value{
A \code{deposit} class (R6 class)

invisibly returns self

dataframe of remote csv file.

String. Path to download location.

data.frame from \code{\link[readr:read_delim]{readr::read_csv()}}

String. cleaned zenodo id

metadata in the specificed format (invisible)

List. Named list of data frames with the following fields:
\itemize{
\item name = field name
\item type = field type
\item description = characterization of the field
}
}
\description{
An R6 client for getting virion data and metadata.
}
\examples{
\dontrun{
# make a client
virion_deposit <- deposit$new()

# methods
virion_deposit$set_working_version("latest")
virion_deposit$working_files

# load remote data directly into your session
virion_data  <- virion_deposit$load_remote_csv_file(file_key = "virion.csv.gz")

# download remote data to disk - defaults to working version
virion_deposit$download_versioned_data()

}

## ------------------------------------------------
## Method `deposit$new`
## ------------------------------------------------

\dontrun{
virion_deposit <- deposit$new()
}

## ------------------------------------------------
## Method `deposit$set_working_version`
## ------------------------------------------------


\dontrun{
# set the working version to the latest version of the deposit
virion_deposit$set_working_version("latest")

# set the working versino to an arbitrary deposit
virion_deposit$all_versions
virion_deposit$set_working_version("19502921")
}


## ------------------------------------------------
## Method `deposit$load_remote_csv_file`
## ------------------------------------------------

\dontrun{
## see the working files
virion_deposit$working_files
virion_data <- virion_deposit$load_remote_csv_file("virion.csv.gz")

}

## ------------------------------------------------
## Method `deposit$download_versioned_data`
## ------------------------------------------------

\dontrun{
# download files for the working version
virion_deposit$download_versioned_data()

# download files for the latest version
virion_deposit$download_versioned_data("latest")

# download files for some zenodo id
virion_deposit$download_versioned_data("19502921")
}


## ------------------------------------------------
## Method `deposit$load_local_csv_file`
## ------------------------------------------------

\dontrun{

virion <- deposit$new()
virion$set_working_version()
virion$working_files$file_key
virion$load_local_csv_file("virion.csv.gz")
}


## ------------------------------------------------
## Method `deposit$check_zenodo_id`
## ------------------------------------------------


\dontrun{
virion_deposit$check_zenodo_id("latest")
}

## ------------------------------------------------
## Method `deposit$export_metadata`
## ------------------------------------------------


\dontrun{
 virion_deposit$export_metadata("working")
}


## ------------------------------------------------
## Method `deposit$get_data_dictionary`
## ------------------------------------------------

\dontrun{
virion <- deposit$new()
virion$set_working_version()
virion$get_data_dictionary()

}

}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{parent_id}}{String. Identifier for a Zenodo deposit with multiple versions.}

\item{\code{all_versions}}{Character. All zenodo ids for versions of the deposit.}

\item{\code{summary_df}}{Data frame. A table of all zenodo ids for versions of the deposit.}

\item{\code{latest_version}}{String. Identifier for the latest version of the zenodo deposit.}

\item{\code{working_version}}{String. Identifier for the version of the zenodo deposit you are working with.}

\item{\code{working_bibtex}}{String. Bibtex for the version of the zenodo deposit you are working with.}

\item{\code{working_citation}}{String. Citation for the version of the zenodo deposit you are working with.}

\item{\code{working_files}}{Data frame. Working files consist of key (file name) and url (url to file on zenodo)}

\item{\code{working_json}}{list. Metadata for the working version.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-deposit-new}{\code{deposit$new()}}
\item \href{#method-deposit-set_working_version}{\code{deposit$set_working_version()}}
\item \href{#method-deposit-load_remote_csv_file}{\code{deposit$load_remote_csv_file()}}
\item \href{#method-deposit-download_versioned_data}{\code{deposit$download_versioned_data()}}
\item \href{#method-deposit-load_local_csv_file}{\code{deposit$load_local_csv_file()}}
\item \href{#method-deposit-check_zenodo_id}{\code{deposit$check_zenodo_id()}}
\item \href{#method-deposit-export_metadata}{\code{deposit$export_metadata()}}
\item \href{#method-deposit-get_data_dictionary}{\code{deposit$get_data_dictionary()}}
\item \href{#method-deposit-clone}{\code{deposit$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-deposit-new"></a>}}
\if{latex}{\out{\hypertarget{method-deposit-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{deposit} object, as an \link[R6:R6-package]{R6}
client. This object is connected to the zenodo deposit for virion
and is the gateway for accessing different versions of the data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{deposit$new(parent_id = "15643003")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{parent_id}}{String. Identifier for a Zenodo deposit with multiple versions.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{deposit} object
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
virion_deposit <- deposit$new()
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-deposit-set_working_version"></a>}}
\if{latex}{\out{\hypertarget{method-deposit-set_working_version}{}}}
\subsection{Method \code{set_working_version()}}{
Set the version of the data you would
like to work with. Note that the structure of Virion
data may change through time.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{deposit$set_working_version(
  zenodo_id = "latest",
  style = "apa",
  print_bibtex = TRUE,
  print_citation = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{zenodo_id}}{String. A zenodo id or latest.}

\item{\code{style}}{String. The citation style to be used.}

\item{\code{print_bibtex}}{Logical. Should the bibtex for the working version of the deposit be printed when setting the working version?}

\item{\code{print_citation}}{Logical. Should the citation for the working version of the deposit be printed?}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{
\dontrun{
# set the working version to the latest version of the deposit
virion_deposit$set_working_version("latest")

# set the working versino to an arbitrary deposit
virion_deposit$all_versions
virion_deposit$set_working_version("19502921")
}

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-deposit-load_remote_csv_file"></a>}}
\if{latex}{\out{\hypertarget{method-deposit-load_remote_csv_file}{}}}
\subsection{Method \code{load_remote_csv_file()}}{
Load a remote csv listed in the working files field
directly into the environment.

Can handle either csv or csv.gz files. This function
downloads a temporary file to your machine, reads it,
then returns a dataframe.

The local file cache is removed when your R session
restarts.

Consider using the \code{download_versioned_data} to
save a persistent version of the virion dataset on
your machine.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{deposit$load_remote_csv_file(file_key, refresh = FALSE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file_key}}{String. A file key as seen in the \code{working_files} field}

\item{\code{refresh}}{Logical. Should files be re-downloaded.}

\item{\code{...}}{Additional arguments to pass to \code{\link[readr:read_delim]{readr::read_csv()}}}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
## see the working files
virion_deposit$working_files
virion_data <- virion_deposit$load_remote_csv_file("virion.csv.gz")

}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-deposit-download_versioned_data"></a>}}
\if{latex}{\out{\hypertarget{method-deposit-download_versioned_data}{}}}
\subsection{Method \code{download_versioned_data()}}{
Download versioned data

Downloads all the files for a particular version of
Virion to \code{dir/zenodo_id}.

This function does not change the working version!
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{deposit$download_versioned_data(
  zenodo_id = "working",
  dir = "outputs",
  refresh = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{zenodo_id}}{String. A zenodo id, working or latest}

\item{\code{dir}}{String. Storage location for files}

\item{\code{refresh}}{Logical. Should files be re-downloaded.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
# download files for the working version
virion_deposit$download_versioned_data()

# download files for the latest version
virion_deposit$download_versioned_data("latest")

# download files for some zenodo id
virion_deposit$download_versioned_data("19502921")
}

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-deposit-load_local_csv_file"></a>}}
\if{latex}{\out{\hypertarget{method-deposit-load_local_csv_file}{}}}
\subsection{Method \code{load_local_csv_file()}}{
Load a local version of the csv.

Assumes you would like to load the file from
the current working version.

If the csv file is not present, a persistent version
of the file will be downloaded to your machine.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{deposit$load_local_csv_file(file_key, refresh, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file_key}}{string. Name of file from \code{working_files}}

\item{\code{refresh}}{logical. Should the file be re-downloaded?}

\item{\code{...}}{Additional parameters to pass to \code{\link[readr:read_delim]{readr::read_csv()}}}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{

virion <- deposit$new()
virion$set_working_version()
virion$working_files$file_key
virion$load_local_csv_file("virion.csv.gz")
}

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-deposit-check_zenodo_id"></a>}}
\if{latex}{\out{\hypertarget{method-deposit-check_zenodo_id}{}}}
\subsection{Method \code{check_zenodo_id()}}{
Check the zenodo ID

Look for key words like working or latest and convert
them to integer ids OR pass the zenodo id directly
to \code{sanitize_id()}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{deposit$check_zenodo_id(zenodo_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{zenodo_id}}{String. One of working, latest, or a zenodo id}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{
\dontrun{
virion_deposit$check_zenodo_id("latest")
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-deposit-export_metadata"></a>}}
\if{latex}{\out{\hypertarget{method-deposit-export_metadata}{}}}
\subsection{Method \code{export_metadata()}}{
Export metadata for a version of the deposit.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{deposit$export_metadata(zenodo_id, format, verbose)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{zenodo_id}}{String. One of "latest","working", or a zenodo id}

\item{\code{format}}{String. One of "json", "json-ld","csl","datacite-json","datacite-xml", "dublincore","marcxml","bibtex","geojson","dcat-ap","codemeta", "cff"}

\item{\code{verbose}}{Logical. Print the metadata}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{
\dontrun{
 virion_deposit$export_metadata("working")
}

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-deposit-get_data_dictionary"></a>}}
\if{latex}{\out{\hypertarget{method-deposit-get_data_dictionary}{}}}
\subsection{Method \code{get_data_dictionary()}}{
Create a list of data dictionaries.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{deposit$get_data_dictionary(
  file_key = "datapackage.json",
  dir = tempdir(),
  refresh
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file_key}}{string. Name of file from \code{working_files}}

\item{\code{dir}}{string. Name of directory}

\item{\code{refresh}}{refresh Logical. Should files be re-downloaded.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
virion <- deposit$new()
virion$set_working_version()
virion$get_data_dictionary()

}

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-deposit-clone"></a>}}
\if{latex}{\out{\hypertarget{method-deposit-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{deposit$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
