% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zenodo_versions.R
\name{deposit_summary}
\alias{deposit_summary}
\title{Simple summary data frame for the deposit}
\usage{
deposit_summary()
}
\value{
data frame. Simplified deposit metadata with the following fields:
\itemize{
\item id = zenodo id
\item latest_version = True or False indicating if that record is the latest version of the dataset.
\item publication_date = YMD the item was published to zenodo
\item doi_url = digital object identifier for that version of the record.
}
}
\description{
Simple summary data frame for the deposit
}
\examples{
deposit_summary()

}
