% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_json.R
\name{download_refresh}
\alias{download_refresh}
\title{Download or redownload data}
\usage{
download_refresh(file_url, local, refresh)
}
\arguments{
\item{file_url}{String. URL for file you'd like to download}

\item{local}{String. A character string with the name where the downloaded file is saved. Tilde-expansion is performed.}

\item{refresh}{Logical. Should saved files be refreshed.}
}
\value{
String. Local File path
}
\description{
Downloads a file or re-downloads a file to refresh the cached files.
}
