% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virion_class.R
\name{dummy_imports}
\alias{dummy_imports}
\title{Dummy function to pass RMD check}
\usage{
dummy_imports()
}
\value{
Character string explaining why this function exists.
}
\description{
Apparently a known issue in R6 as the class is not created until install.
}
