% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribution.R
\name{export_deposit_bibtex}
\alias{export_deposit_bibtex}
\title{Get bibtex entry for a deposit.}
\usage{
export_deposit_bibtex(zenodo_id = the$working_version, verbose = TRUE)
}
\arguments{
\item{zenodo_id}{String. ID for a Zenodo deposit. Should correspond to the version of a deposit.}

\item{verbose}{Logical. Print the formatted metadata?}
}
\value{
Character. bibtex formatted entry
}
\description{
Bibtex specific wrapper for export_deposit_metadata
}
\examples{
\dontrun{
export_deposit_bibtex(zenodo_id = "15692263" ) |>
  writeLines(con = "outputs/citation.bib")
}
}
