% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribution.R
\name{export_deposit_metadata}
\alias{export_deposit_metadata}
\title{Export Deposit Metadata in various formats}
\usage{
export_deposit_metadata(
  zenodo_id = the$working_version,
  format,
  verbose = TRUE
)
}
\arguments{
\item{zenodo_id}{String. ID for a Zenodo deposit. Should correspond to the version of a deposit.}

\item{format}{String. File format for export. One of "json", "json-ld","csl","datacite-json","datacite-xml",
"dublincore","marcxml","bibtex","geojson","dcat-ap","codemeta", or "cff"}

\item{verbose}{Logical. Print the formatted metadata?}
}
\value{
Character. Returns text in selected format.
}
\description{
Export Deposit Metadata in various formats
}
\examples{
 \dontrun{
export_deposit_metadata(zenodo_id = "15692263","json-ld" ) |>
  writeLines(con = "outputs/citation.jsonld")
}
}
