% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zenodo_versions.R
\name{get_all_versions}
\alias{get_all_versions}
\title{Get all deposit versions}
\usage{
get_all_versions(url)
}
\arguments{
\item{url}{Character. URL for a version of the deposit.}
}
\value{
Data.frame. Metadata for all deposit version
}
\description{
Recursively traverse a deposit link to get all versions of a deposit.
}
\examples{

\dontrun{

self_link <- "https://zenodo.org/api/records/21232661/versions?page=1&size=25&sort=version"
get_all_versions(self_link)
}

}
