% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data_dictionary.R
\name{get_data_dictionary}
\alias{get_data_dictionary}
\title{Get Data Dictionary for a data Package}
\usage{
get_data_dictionary(datapackage_json)
}
\arguments{
\item{datapackage_json}{Character. Path to datapackage.json file}
}
\value{
List. Named list of data frames with the following fields:
* name = field name
* type = field type
* description = characterization of the field
}
\description{
Creates a list of data frames that constitutes the data dictionary.
}
\examples{

\dontrun{
get_data_dictionary(datapackage_json = "data/1235600/datapackage.json")
}

}
