% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data_dictionary.R
\name{get_individual_dictionary}
\alias{get_individual_dictionary}
\title{Get the dictionary for a single resource}
\usage{
get_individual_dictionary(x)
}
\arguments{
\item{x}{List. Resource object from the data package.}
}
\value{
dataframe with three fields.
}
\description{
Get the dictionary for a single resource
}
