% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_json.R
\name{get_json}
\alias{get_json}
\title{Get json}
\usage{
get_json(url)
}
\arguments{
\item{url}{String. A url fed to \link{jsonlite::fromJSON}}
}
\value{
List. JSON as a list
}
\description{
Get json
}
\examples{
\dontrun{
get_json("https://www.example.com/example.json")
}

}
