% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data_dictionary.R
\name{get_resource_name}
\alias{get_resource_name}
\title{Get the name of a resource}
\usage{
get_resource_name(x)
}
\arguments{
\item{x}{data frame with name field}
}
\value{
character. Name of resources
}
\description{
Get the name of a resource
}
