% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribution.R
\name{get_version_bibtex}
\alias{get_version_bibtex}
\title{Get bibtex for citaiton}
\usage{
get_version_bibtex(zenodo_id = the$working_version, verbose = TRUE)
}
\arguments{
\item{zenodo_id}{Character}

\item{verbose}{Logical}
}
\value{
Character. bibtex formatted entry
}
\description{
Get bibtex for citaiton
}
\examples{
\dontrun{
get_version_bibtex(zenodo_id ) |>
  writeLines(con = "outputs/citation.bib")
}
}
