% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zenodo_versions.R
\name{get_versioned_data}
\alias{get_versioned_data}
\title{Get a specific version of the data}
\usage{
get_versioned_data(
  version = "latest",
  style = "apa",
  dir_path,
  refresh_deposits_versions = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{version}{Character. identifier for a version e.g. "15643003" or "latest"}

\item{style}{Charater. Character. One of "havard-cite-them-right",
"apa",
"modern-language-association",
"vancouver",
"chicago-fullnote-bibliography", or
"ieee"}

\item{dir_path}{Character. Path to folder where files should be downloaded.}

\item{refresh_deposits_versions}{Logical. Should the function check for new
versions of the deposit?}

\item{verbose}{Logical. Include print statements?}

\item{...}{Additional arguments to pass to `download_deposit_version`}
}
\value{
Character. path to versioned data.
}
\description{
This is a convenience function that downloads a specific version of the
 data and returns attribution information.
}
\details{
If you would like to download a bibtex entry use `export_deposit_bibtex`. For
 other metadata types (CSL, CFF, JSON-LD, etcetera) use `export_deposit_metadata`.
}
