% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zenodo_versions.R
\name{list_deposit_versions}
\alias{list_deposit_versions}
\title{Get Versions of a Deposit on Zenodo}
\usage{
list_deposit_versions(parent_id = "15643003")
}
\arguments{
\item{parent_id}{String. Identifier for a Zenodo deposit with multiple versions.}
}
\value{
Data frame. Invisible. The data frame contains the Zenodo id for each version of
the deposit, as well as the version name, and logical field called latest that
indicates if this is the latest version.
}
\description{
This function gets the metadata for all the versions of a deposit associated
with a parent id. The parent id is used to identify a set of works that are
different versions of the same work. The parent id is provided by the
Zenodo API. To find the parent ID, download a JSON representation of the
deposit (export to json on the webpage or use `export_deposit_metadata`),
there will be an attribute called parent that looks like
"https://zenodo.org/api/records/15020049".
The 8 digit string at the end of the url is the parent id.
}
\details{
This function also sets package variables
}
