% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_json.R
\name{make_url}
\alias{make_url}
\title{Make Zenodo URLs}
\usage{
make_url(base_url = "https://zenodo.org/api/records/\%s", id)
}
\arguments{
\item{id}{}
}
\value{
String. URL for zenodo api.
}
\description{
Appends a zenodo id to the zenodo api url.
}
\examples{
make_url(id = "some_id")

}
