% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zenodo_versions.R
\name{sanitize_id}
\alias{sanitize_id}
\title{Sanitize a zenodo id}
\usage{
sanitize_id(zenodo_id)
}
\arguments{
\item{zenodo_id}{String. A zenodo id.}
}
\value{
String. A cleaned zenodo id.
}
\description{
Remove any white space and check that it conforms to the zenodo id pattern.
}
\examples{
sanitize_id(" 2948598")

}
