% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zenodo_versions.R
\name{sanitize_version}
\alias{sanitize_version}
\title{Check version ids}
\usage{
sanitize_version(version)
}
\arguments{
\item{version}{Character. Version identifier.}
}
\value{
Character. Version identifier from Zenodo.
}
\description{
Checks that version ids are properly formatted. IDs should
 either be integers ("15643003") OR "latest"
}
\examples{

sanitize_version("latest")
sanitize_version(" 15643003")

}
