% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virionData-package.R
\docType{package}
\name{virionData-package}
\alias{virionData}
\alias{virionData-package}
\title{virionData: Access Versioned VIRION Data}
\description{
This package provides an ergonomic interface for accessing the VIRION dataset (https://doi.org/10.5281/zenodo.15643003).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/viralemergence/virionData}
  \item \url{https://viralemergence.github.io/virionData/}
  \item Report bugs at \url{https://github.com/viralemergence/virionData/issues}
}

}
\author{
\strong{Maintainer}: Collin Schwantes \email{collin@viralemergence.org} (\href{https://orcid.org/0000-0002-9882-941X}{ORCID})

}
\keyword{internal}
