% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zenodo_versions.R
\name{batch_download_deposit_versions}
\alias{batch_download_deposit_versions}
\title{Batch download deposit versions}
\usage{
batch_download_deposit_versions(zenodo_ids = "all", dir_path)
}
\arguments{
\item{zenodo_ids}{Character. Either a vector of zenodo ids or "all"}

\item{dir_path}{Character. Path to folder where files should be downloaded.}
}
\value{
List of download locations.
}
\description{
This is `download_deposit_version` wrapped in a `purr::map` call.
}
\examples{
\dontrun{
# get all deposit versions
batch_download_deposit_versions(dir_path = "outputs")

# get select versions
batch_download_deposit_versions(zenodo_ids = c(15677137", "15643004"), dir_path = "outputs)
}

}
