% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zenodo_versions.R
\name{get_versioned_data}
\alias{get_versioned_data}
\title{Get a specific version of the data}
\usage{
get_versioned_data(
  version = "latest",
  style = "apa",
  dir_path,
  refresh_deposits_versions = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{version}{Character. identifier for a version e.g. "15643003" or "latest"}

\item{style}{Charater.}

\item{dir_path}{Character.}

\item{refresh_deposits_versions}{Logical.}

\item{verbose}{Logical.}
}
\value{
Character. path to versioned data.
}
\description{
Get a specific version of the data
}
